requirements.txt
cv2
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
Pillow==7.1.2
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
PyQt5-stubs==5.14.2.2
python-dateutil==2.8.1
six==1.15.0
